% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_datasets.R
\name{combine_datasets}
\alias{combine_datasets}
\title{Combine extracted simulation datasets}
\usage{
combine_datasets(inpaths, vars, params, merge_factor, merge_levels)
}
\arguments{
\item{inpaths}{Where the simulation outputs are to be found. Need several of them, of course.}

\item{vars}{What variables through time to extract from simulation output data files?}

\item{params}{What parameter values to record for each simulation?}

\item{merge_factor}{The name of the column to add to the master data frame; the factor that differs between the different datasets that are combined.}

\item{merge_levels}{The levels of that factor. Should be of the same length as \code{inpaths}.}
}
\value{
A data frame with individual simulationa in rows, and in columns the variables of interest at each time point, as well as the values of the selected parameters.
}
\description{
This function extracts data from different input paths and combine them. It is a wrapper around \code{extract_from_simulations}.
}
