% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_from_simulations.R
\name{extract_from_simulations}
\alias{extract_from_simulations}
\title{Extract data from simulations}
\usage{
extract_from_simulations(inpath, vars, params)
}
\arguments{
\item{inpath}{Where the simulation outputs are to be found.}

\item{vars}{What variables through time to extract from simulation output data files?}

\item{params}{What parameter values to record for each simulation?}
}
\value{
A data frame with individual simulationa in rows, and in columns the variables of interest at each time point, as well as the values of the selected parameters.
}
\description{
This function loops through the simulation output folders and reads in the information from the variables and parameter values of interest. It writes all this into a large data frame that can then be saved as CSV. Much lighter to analyze than going through all simulations folders every time.
}
