% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_parameters.R
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract parameters of the simulations}
\usage{
extract_parameters(datapath, outpath,
  whichparams = c("habitat_asymmetry", "sel_coeff_ecol"))
}
\arguments{
\item{datapath}{Path to where the simulation folders are.}

\item{outpath}{Where to save the output data frame. Optional.}

\item{whichparams}{What parameters to extract?}
}
\value{
A data frame with the parameter values for each simulation.
}
\description{
Extract parameters of the simulations
}
