% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trajectories.R
\name{extract_trajectories}
\alias{extract_trajectories}
\title{Extract trajectories in speciation cube}
\usage{
extract_trajectories(datapath, outpath)
}
\arguments{
\item{datapath}{Path to where the simulation folders are.}

\item{outpath}{Where to save the output data frame.}
}
\value{
A data frame with one column for time and three columns per simulation containing coordinates in speciation cube. The data frame is saved as a CSV file in outpath.
}
\description{
Extract trajectories in speciation cube
}
