% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_completed.R
\name{find_completed}
\alias{find_completed}
\title{Find completed events}
\usage{
find_completed(speciation_cube_data, dim, threshold, n)
}
\arguments{
\item{speciation_cube_data}{A data frame containing trajectories in speciation cube for all simulations (in rows).}

\item{dim}{The variable to use as a criterion for complete speciation.}

\item{threshold}{The threshold above which the value of the criterion variable means complete speciation}

\item{n}{How many of the last time points to average?}
}
\value{
A subset of \code{speciation_cube_data}.
}
\description{
Function to find sets of simulations reaching a certain threshold in certain dimensions.
}
