% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_simulations.R
\name{find_simulations}
\alias{find_simulations}
\title{Find simulations}
\usage{
find_simulations(speciation_cube_data, path = ".",
  pars = list(sel_coeff_ecol = 0.8, habitat_asymmetry = 1))
}
\arguments{
\item{speciation_cube_data}{A data frame containing trajectories in speciation cube for all simulations (in rows).}

\item{path}{If \code{speciation_cube_data} not provided, the path to where the simulation data are to be found.}

\item{pars}{A list of parameter values. The names in the list should be the names of the parameters as they appear in the parameters.txt file.}
}
\value{
A logical vector indicated simulations that match the search.
}
\description{
Find simulations
}
