% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_complete.R
\name{is_complete}
\alias{is_complete}
\title{Is a speciation event complete?}
\usage{
is_complete(folder, ri.threshold = 0.9)
}
\arguments{
\item{folder}{A string. The path to the folder of a given simulation.}

\item{ri.threshold}{Numeric. The threshold value of final mating isolation above which speciation is complete.}
}
\value{
A vector containing the habitat asymmetry, selection coefficient and whether speciation is complete (0 or 1).
}
\description{
Is a speciation event complete?
}
