% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fst_map.R
\name{plot_fst_map}
\alias{plot_fst_map}
\title{Fst scan through time}
\usage{
plot_fst_map(folder)
}
\arguments{
\item{folder}{A string. The path to the folder of a given simulation.}
}
\description{
Fst scan through time
}
