% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fst_scan.R
\name{plot_fst_scan}
\alias{plot_fst_scan}
\title{Is a speciation event complete?}
\usage{
plot_fst_scan(folder, time = 5e+05)
}
\arguments{
\item{folder}{A string. The path to the folder of a given simulation.}

\item{time}{Integer. The time step at which to plot the Fst scan.}
}
\description{
Is a speciation event complete?
}
