% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_scan.R
\name{plot_genome_scan}
\alias{plot_genome_scan}
\title{Plot genome scan}
\usage{
plot_genome_scan(path = ".", var, time = 5e+05)
}
\arguments{
\item{path}{Where the data for the given simulation are to be found.}

\item{var}{The variable to plot along the genome.}

\item{time}{What time point?}
}
\description{
This functions plots a scan of a given variable along the genome.
}
