% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_scan_through_time.R
\name{plot_genome_scan_through_time}
\alias{plot_genome_scan_through_time}
\title{Plot genome scan through time}
\usage{
plot_genome_scan_through_time(path = ".", var)
}
\arguments{
\item{path}{Where the data for the given simulation are to be found.}

\item{var}{The variable to plot along the genome.}
}
\description{
This functions plots a heatmap of a given variable along the genome, through time.
}
