% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_projected_trajectories.R
\name{plot_projected_trajectories}
\alias{plot_projected_trajectories}
\title{Plot simulations}
\usage{
plot_projected_trajectories(speciation_cube_data, vars, colvar,
  show_legend = T, xlim, ylim, confint, col, add = F)
}
\arguments{
\item{speciation_cube_data}{A data frame containing coordinates along the three isolation axes (spatial, ecological and mating) through time. Individual simulations are in rows.}

\item{vars}{The two variables along which to plot the simulations.}

\item{colvar}{Optional. The name of the variable to use to set colors according to.}

\item{show_legend}{Optional. Whether to show the legend.}

\item{xlim, ylim}{Ranges of values to plot.}

\item{confint}{Confidence interval probabilities. If specified, the confidence interval is plotted instead of every single line. To plot mean and standard error instead of a confidence interval, provide this argument with an object of length different than 2 (which is what the function expects as boundaries of the confidence interval).}

\item{col}{Color to plot the lines or the confidence interval. If specified, overrides the coloration per parameter value, if any.}

\item{add}{Whether to append to a previous plot. Defaults to FALSE.}
}
\description{
This functions plots the simulations in 2D, along chosen dimensions.
}
