% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_speciation_cube.R
\name{plot_speciation_cube}
\alias{plot_speciation_cube}
\title{Speciation cube}
\usage{
plot_speciation_cube(speciation_cube_data)
}
\arguments{
\item{speciation_cube_data}{A data frame containing coordinates along the three isolation axes (spatial, ecological and mating) through time. Individual simulations are in rows.}
}
\description{
This functions plots the simulations in speciation cube. Avoid using it to visualize directly in RStudio, takes too long. Instead, prefer calling a pdf device before running the function, a view the PDF. Much faster.
}
