% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_speciation_scape.R
\name{plot_speciation_scape}
\alias{plot_speciation_scape}
\title{Speciation scape}
\usage{
plot_speciation_scape(speciation_cube_data, paramspace, dim, threshold,
  n = 1)
}
\arguments{
\item{speciation_cube_data}{A data frame containing trajectories in speciation cube for all simulations (in rows).}

\item{paramspace}{The names of the two parameters to plot against.}

\item{dim}{The variable to use as a criterion for complete speciation.}

\item{threshold}{The threshold above which the value of the criterion variable means complete speciation}

\item{n}{How many of the last time points to average?}
}
\description{
This is a function to visualize the speciation probability landscape in a heatmap across parameter space.
}
