% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_through_time.R
\name{plot_through_time}
\alias{plot_through_time}
\title{Plot simulation metric through time}
\usage{
plot_through_time(data, var, type = 1, colvar, xlim, ylim, confint, cols,
  show_legend = F)
}
\arguments{
\item{data}{A data frame containing coordinates along the three isolation axes (spatial, ecological and mating), or even other variables, through time. Individual simulations are in rows.}

\item{var}{The variable of interest.}

\item{type}{Either 1 -- shows lines for all simulations, 2 -- shows median and specified quantiles, 3 -- shows mean and standard error.}

\item{colvar}{Optional. The variable used to assign different colors to different groups of simulations.}

\item{xlim, ylim}{Plot limits (optional).}

\item{confint}{The quantiles to use as borders of shaded areas (only if type == 2).}

\item{cols}{A vector of colors to use for each group, only if colvar is specified.}

\item{show_legend}{Whether to display the legend. Defaults to FALSE.}
}
\description{
This functions makes plots to track the advancement of the simulation through time.
}
