% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectories.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{Plot trajectories in speciation cube}
\usage{
plot_trajectories(trajectories, smoothing = T, nbins = 30,
  ri.threshold = 0.9)
}
\arguments{
\item{trajectories}{A list of data frames. Each data frame has three columns, one for each dimension of the cube. Typically the output of `extract_trajectories`.}

\item{smoothing}{Logical. `TRUE` if you want to average part of the trajectories into bins of given sizes.}

\item{nbins}{Integer. If `smoothing = TRUE`, the number of bins to calculate averages for when smoothing.}

\item{ri.threshold}{Numeric. If `smoothing = TRUE`, this argument specifies the value of mating isolation over which determines the part of the trajectory that has to be smoothed. A trajectory is smoothed onwards from the time at which it irreversibly goes above `ri.threshold` in the course of the simulation. This makes the plot lighter while keeping the most informative part of each trajectory intact (i.e. the way towards reproductive isolation).}
}
\description{
Plot trajectories in speciation cube
}
