% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothen_bins.R
\name{smoothen_bins}
\alias{smoothen_bins}
\title{Smoothen a trajectory}
\usage{
smoothen_bins(df, nbins)
}
\arguments{
\item{df}{A data frame with numeric columns.}

\item{nbins}{Integer. The number of bins to calculate averages for when smoothing.}
}
\value{
A matrix with the same number of columns as `df` and `nbins` columns.
}
\description{
Smoothen a trajectory
}
