% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_completed.R
\name{subset_completed}
\alias{subset_completed}
\title{Subset completed events}
\usage{
subset_completed(speciation_cube_data, dim, threshold, n)
}
\arguments{
\item{speciation_cube_data}{A data frame containing trajectories in speciation cube for all simulations (in rows).}

\item{dim}{The variable to use as a criterion for complete speciation.}

\item{threshold}{The threshold above which the value of the criterion variable means complete speciation}

\item{n}{How many of the last time points to average?}

\item{paramspace}{The names of the two parameters to plot against.}
}
\value{
A subset of \code{speciation_cube_data}.
}
\description{
Subset completed events
}
