% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trajectories.R
\name{extract_trajectories}
\alias{extract_trajectories}
\title{Extract trajectories in speciation cube}
\usage{
extract_trajectories(dir, is.complete)
}
\arguments{
\item{dir}{A string. The path to the folders containing the simulations.}

\item{is.complete}{A data frame. Contains information about habitat asymmetry, selection coefficient and whether speciation was complete or not (see argument `ri.threshold`) for each simulation. Typically the output of `find_completed`.}
}
\value{
A list of data frames, one for each simulation. Each data frame contains three columns, one for each dimension of the speciation cube.
}
\description{
Extract trajectories in speciation cube
}
