% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_completed.R
\name{find_completed}
\alias{find_completed}
\title{Find completed speciation events}
\usage{
find_completed(dir = ".", ri.threshold = 0.9)
}
\arguments{
\item{dir}{A string. The path to the folders containing the simulations.}

\item{ri.threshold}{Numeric. The threshold value of final mating isolation above which speciation is complete.}
}
\value{
A data frame containing information about habitat asymmetry, selection coefficient and whether speciation is complete, for each simulation.
}
\description{
Find completed speciation events
}
