% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_projected_trajectories.R
\name{plot_projected_trajectories}
\alias{plot_projected_trajectories}
\title{Plot simulations}
\usage{
plot_projected_trajectories(speciation_cube_data, vars, colvar)
}
\arguments{
\item{speciation_cube_data}{A data frame containing coordinates along the three isolation axes (spatial, ecological and mating) through time. Individual simulations are in rows.}

\item{vars}{The two variables along which to plot the simulations.}

\item{colvar}{Optional. The name of the variable to use to set colors according to.}
}
\description{
This functions plots the simulations in 2D, along chosen dimensions.
}
